Separated profileAvatar and rendering correctly
