+++ client/src/Profile.jsx
@@ -1,18 +1,13 @@
 import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import Avatar from 'react-avatar-edit';
 import TextField from '@material-ui/core/TextField';
 import MenuItem from '@material-ui/core/MenuItem';
 import Button from '@material-ui/core/Button';
-import { saveProfile, uploadPhoto } from './actionCreators';
+import { saveProfile } from './actionCreators';
+import ProfileAvatar from './ProfileAvatar';
 
 const styles = {
-  avatarContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-    marginTop: 50,
-    marginBottom: 50,
-  },
   textField: {
     width: 500,
   },
@@ -24,7 +19,6 @@ function Profile({
   defaultInterests,
   defaultDietRestrictions,
   dispatchSaveProfile,
-  dispatchUploadPhoto,
 }) {
   const [firstName, setFirstName] = useState(defaultFirstName);
   const [lastName, setLastName] = useState(defaultLastName);
@@ -35,8 +29,6 @@ function Profile({
     defaultDietRestrictions,
   );
   const [dietOptionOther, setDietOptionOther] = useState(false);
-  const [preview, setPreview] = useState(null);
-  const [src] = useState('');
   const dietOptions = [
     'None - I eat anything & everything!',
     'Vegan',
@@ -45,26 +37,6 @@ function Profile({
     'Other',
   ];
   const specifyOtherDiet = React.createRef();
-
-  function onCrop(currView) {
-    setPreview(currView);
-    console.log('this is src', src);
-  }
-
-  function onClose() {
-    setPreview(null);
-  }
-
-  function onBeforeFileLoad(file) {
-    if (file.target.files[0].size > 300000) {
-      alert('File is too big!');
-      file.target.value = '';
-    }
-  }
-
-  function onFileLoad(file) {
-    dispatchUploadPhoto(file);
-  }
 
   function handleDietOption(e) {
     if (e.target.value === 'Other') {
@@ -105,7 +77,7 @@ function Profile({
         lastName,
         password,
         interests,
-        dietRestrictions
+        dietRestrictions,
       );
       console.log('this is userChanges', userChanges);
 
@@ -118,18 +90,7 @@ function Profile({
 
   return (
     <div>
-      <div style={styles.avatarContainer}>
-        <Avatar
-          width={295}
-          height={200}
-          onCrop={onCrop}
-          onClose={onClose}
-          onBeforeFileLoad={onBeforeFileLoad}
-          src={src}
-          onFileLoad={onFileLoad}
-        />
-        <img src={preview} alt="Preview" />
-      </div>
+      <ProfileAvatar />
       <form onSubmit={e => onSubmit(e)}>
         <TextField
           style={styles.textField}
@@ -210,15 +171,31 @@ function Profile({
   );
 }
 
+Profile.propTypes = {
+  defaultFirstName: PropTypes.string,
+  defaultLastName: PropTypes.string,
+  defaultInterests: PropTypes.arrayOf(PropTypes.string),
+  defaultDietRestrictions: PropTypes.string,
+  dispatchSaveProfile: PropTypes.func,
+};
+
+Profile.defaultProps = {
+  defaultFirstName: '',
+  defaultLastName: '',
+  defaultInterests: [],
+  defaultDietRestrictions: '',
+  dispatchSaveProfile: () => {},
+};
+
 const mapStateToProps = ({ profile }) => ({
   defaultFirstName: profile.firstName,
   defaultLastName: profile.lastName,
   defaultInterests: profile.interests,
   defaultDietRestrictions: profile.dietRestrictions,
 });
+
 const mapDispatchToProps = {
   dispatchSaveProfile: saveProfile,
-  dispatchUploadPhoto: uploadPhoto,
 };
 
 export default connect(
